* Increase constness of CMobileCAI::SetGoal (needed for branches/tvo3) * Merge from branches/tvo3:   * Infolog writes to file before line breaking   * Fix gcc warning in BFGroundDrawer
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@1418 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Map/SMF/BFGroundDrawer.cpp
@@ -105,7 +105,7 @@ void CBFGroundDrawer::Draw(bool drawWaterReflection,bool drawUnitReflection,unsi
 	int hmapx=mapx>>1;
 	int mapy=gs->mapy+1;
 
-	int neededLod=(gu->viewRange/8/viewRadius*2);
+	int neededLod=int(gu->viewRange/8/viewRadius*2);
 	UpdateCamRestraints();
 
 	invMapSizeX=1.0/gs->mapx;
